feat: Set debug log level on verbosity > 2 (#239)
## Description
We have debug logging calls in a few places but we don't set the logger
to debug level anywhere. This sets all loggers to debug level when
verbosity > 2

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***