fix(grib output): Respect `write_initial_state` option (#238)
## Description
Grib output was not respecting this option and always writing the
initial state. This check exists in the base function, but was lost in
the derived function.

## What issue or task does this change relate to?
closes #237

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***