feat: inference with external graph. (#216)
## Description
We add functionality so that in inference the model can use another
graph than the one it was trained with. In this first implementation the
graph has to be provided as a file on disk.

This PR adds a new runner `runner: external_graph` that is an extension
of the default runner. The code is based on a similar feature in
bris-inference:
https://github.com/metno/bris-inference/blob/main/bris/checkpoint.py#L185

The runner can be selected and set in the config as follows:

```yaml
runner:
  external_graph:
    graph: path/to/graph.pt
```
For further options for the runner please consult the documentation.

<!-- Provide a brief summary of the changes introduced in this pull
request. -->

## Type of Change

-   [ ] Bug fix (non-breaking change which fixes an issue)
-   [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
-   [x] Documentation update

## Issue Number

<!-- Link the Issue number this change addresses, ideally in one of the
"magic format" such as Closes #XYZ -->
Closes #215 .
<!-- Alternatively, explain the motivation behind the changes and the
context in which they are being made. -->

## Code Compatibility

-   [x] I have performed a self-review of my code

### Code Performance and Testing

- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I ran the [complete Pytest
test](https://anemoi.readthedocs.io/projects/training/en/latest/dev/testing.html)
suite locally, and they pass

<!-- In case this affects the model sharding or other specific
components please describe these here. -->

### Dependencies

- [x] I have ensured that the code is still pip-installable after the
changes and runs
- [ ] I have tested that new dependencies themselves are
pip-installable.

<!-- List any new dependencies that are required for this change and the
justification to add them. -->

### Documentation

-   [ ] My code follows the style guidelines of this project
- [x] I have updated the documentation and docstrings to reflect the
changes
- [x] I have added comments to my code, particularly in
hard-to-understand areas

<!-- Describe any major updates to the documentation -->

## Additional Notes

<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->


<!-- readthedocs-preview anemoi-inference start -->
----
📚 Documentation preview 📚:
https://anemoi-inference--216.org.readthedocs.build/en/216/

<!-- readthedocs-preview anemoi-inference end -->

---------

Co-authored-by: Gert Mertes <gert.mertes@ecmwf.int>
Co-authored-by: Francesco Zanetta <zanetta.francesco@gmail.com>