feat: Dynamic parallel runner (run any runner in parallel) (#242)
## Description
Currently, the `ParallelRunner` is hardcoded to inherit from the
`DefaultRunner`. This means you cannot run another runner in parallel
without creating another parallel runner for it.

This PR implements a dynamic parallel runner: the `ParallelRunner` class
is created with dynamic class inheritance. The user can provide from
which base class they want the `ParallelRunner` to inherit from like
this:

```yaml
   runner:
     parallel:
       base_runner: my-custom-runner
```

The simple syntax of `runner: parallel` still works, this uses the
default runner like before.

The parallel runner code now lives in a mixin. It sits transparently on
top of the base class, forwarding everything to the latter.

I added in a provision so that runners can modify themselves for
parallel operation. In the future, if needed, we could go one step
further and let any Runner provide their own complete
ParallelRunnerMixin.

##  Additional notes ##
This should not be a breaking change, the default behaviour is the same
as before (to inherit from the `DefaultRunner`).

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***


<!-- readthedocs-preview anemoi-inference start -->
----
📚 Documentation preview 📚:
https://anemoi-inference--242.org.readthedocs.build/en/242/

<!-- readthedocs-preview anemoi-inference end -->