fix: apply mask to all sources in cutout input. (#249)
## Description
In the Cutout input file, the mask of the first source is never applied.

## What problem does this change solve?
Previously, the mask of the first source in config.input.cutout was not
applied.
For example with the following config:
```yaml
input:
  cutout:
    global:
        grib: path/to/global.grib
    lam_0:
        grib: path/to/lam_0.grib
```

The cutout mask of `global` is never applied. 
This applies the masks to all sources.

## What issue or task does this change relate to?
Fixes #248

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***