fix: Improve external graph (#247)
## Description
Improvements being made to the external graph runner to allow the
supporting arrays to be updated from the graph.
Additionally, removes anemoi-datasets as a hard dependancy.

## What problem does this change solve?
Enables broader usage of this runner.


***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***


<!-- readthedocs-preview anemoi-inference start -->
----
📚 Documentation preview 📚:
https://anemoi-inference--247.org.readthedocs.build/en/247/

<!-- readthedocs-preview anemoi-inference end -->