feat: temporal interpolation runner (#227)
## Description
Implement the inference for the Interpolator model (without being
coupled to a forecaster)>


Currently `self.create_target_forcings` uses both the `input_state` and
the `input_tensor_torch` - this can be refactored by inferring the shape
of the forcings directly.

I also think the `self.forecast` method should be implemented
differently. It currently expects a `lead_time` argument but this is not
used and perhaps _forecast_ is not the best name for this task and
should be something more general. Perhaps we should change the `run()`
method of the base runner?

It is yet to be implemented for parallel inference and using a
forecasting model to produce the target forcings during inference.

## Type of Change

-   [ ] Bug fix (non-breaking change which fixes an issue)
-   [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
-   [ ] Documentation update

## Issue Number

<!-- Link the Issue number this change addresses, ideally in one of the
"magic format" such as Closes #XYZ -->

<!-- Alternatively, explain the motivation behind the changes and the
context in which they are being made. -->

Since
https://github.com/ecmwf/anemoi-core/commit/ea644ce1c9aef902333d9cbb30bcde0a3746fbcc
introduced functionality for interpolating between two weather states
(e.g given 0h, +6h model outputs (+1h, +2h ..., +5h)). Changes are
required in `anemoi-inference` as well to accommodate for this new
functionality.

## Code Compatibility

-   [x] I have performed a self-review of my code

### Code Performance and Testing

- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I ran the [complete Pytest
test](https://anemoi.readthedocs.io/projects/training/en/latest/dev/testing.html)
suite locally, and they pass

<!-- In case this affects the model sharding or other specific
components please describe these here. -->

### Dependencies

- [x] I have ensured that the code is still pip-installable after the
changes and runs
- [ ] I have tested that new dependencies themselves are
pip-installable.

<!-- List any new dependencies that are required for this change and the
justification to add them. -->

### Documentation

-   [ ] My code follows the style guidelines of this project
- [ ] I have updated the documentation and docstrings to reflect the
changes
- [ ] I have added comments to my code, particularly in
hard-to-understand areas

<!-- Describe any major updates to the documentation -->

## Additional Notes

<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->


<!-- readthedocs-preview anemoi-inference start -->
----
📚 Documentation preview 📚:
https://anemoi-inference--227.org.readthedocs.build/en/227/

<!-- readthedocs-preview anemoi-inference end -->

---------

Co-authored-by: Rilwan Adewoyin <18564167+Rilwan-Adewoyin@users.noreply.github.com>