feat!: Stop accumulating from start of forecast by default (#265)
## Description
To accumulate accumulation fields from the start of the forecast, the
user now has to explicitly add the `accumulate_from_start_of_forecast`
post_processor to their config.

We keep showing a warning in the log just in case (printed at the end of
the run just to make sure people see it). After some time we can remove
this warning, or instead of a warning we show some kind of summary of
the forecast with this information.

## What issue or task does this change relate to?
closes #131

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)