feat: inner-level processors (#260)
## Description
Introduces input-level pre-processors and output-level post-processors.

## What problem does this change solve?
Currently, the pre-processors are created by the runner and they are
used indifferently on all input sources.

This is problematic in the case of a cutout where we might want
different pre-processors for the different sources (e.g. use remove_nans
for the local source and nothing for the global source).

Similarly for post-processors when using the `tee` output.

## What issue or task does this change relate to?
Fixes #251.

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***


<!-- readthedocs-preview anemoi-inference start -->
----
📚 Documentation preview 📚:
https://anemoi-inference--260.org.readthedocs.build/en/260/

<!-- readthedocs-preview anemoi-inference end -->