test: Add integration test framework (#241)
## Description

This PR sets up a minimal version for integration tests. 

1. It reads in metadata from a json and attaches it to a dummy model. 
2. A checkpoint is saved to disk
3. The checkpoint is loaded to run inference and do checks

Todo:
- [x] sanity checks on grib output. 

## What problem does this change solve?
<!-- Describe if it's a bugfix, new feature, doc update, or breaking
change -->

## What issue or task does this change relate to?
Closes #207 

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

---------

Co-authored-by: Gert Mertes <gert.mertes@ecmwf.int>