feat(post-processors): extraction post-processors (#285)
## Description 

This PR adds two new post-processors `extract_mask` and `extract_slice`
(the extraction logic is shared in a common class), which can be used to
extract subsets of points from the state based on boolean masks or
slices respectively. Closes #281.

The post-processors can be used as follows:

```yaml
output:
  printer:
    post_processors:
      - extract_slice: [0, 189700]
```

or 

```yaml
output:
  printer:
    post_processors:
      - extract_mask: mask/in/supporting_arrays or /path/to/mask.npy
```