feat(post-processors): add `assign_mask` post-processor (#287)
## Description
This PR adds a new `assign_mask` post-processor that can be used to
assign the arrays in the current state object to arrays of a larger size
using a boolean array mask. It can be seen as the inverse of
`extract_mask` introduced in #285.

The corresponding output type is now deprecated and will be removed in a
future release.

- [x] added the new post-processor class
- [x] tested locally on GRIB output (as mentioned in
https://github.com/ecmwf/anemoi-inference/pull/285#issuecomment-3132260667)
- [x] added unit tests

---

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)