test: Improve integration test and add RMI LAM (#308)
## Description
This PR adds the RMI LAM model to the integration test. As part of this,
the following was done:

1. Add a helper script to add new models.
2. The mock model now passes prognostic input variables of the last
input time step to the output. Diagnostics are output as all ones. This
allows us to test the full data pipeline and check output values against
the input.
3. Add a new `${s3:}` resolver available in the test config, this
resolves to the model path on S3 where test data is located.
4. The integration test now downloads supporting arrays from S3 (if they
are present in the metadata). Sequential downloading was slower than I
liked (the download itself is fast, but opening the connection takes a
second), using a few threads speeds it up.
5. Add integration test docs
6. Small fix in the unit tests: rename TestingProcessor to
DummyProcessor to avoid pytest trying to pick it up as a test class, and
the simple model was missing prognostics in the input data indices.
7. Add support for multiple input files. Not used in this model but will
be used by MeteoSwiss.

### Helper script
The helper script is run on a real inference checkpoint, trained at a
reduced resolution. See the docs for more explanation.
```shell
$ python anemoi-inference/tests/integration/add_new_model.py -h

usage: add_new_model.py [-h] [--files [FILES ...]] [--overwrite] model checkpoint

Add a new model for integration tests.
Running this script will create a new model directory in tests/integration
with metadata and config files, and upload necessary files to S3.

positional arguments:
  model             Model name. Can only contain alphanumeric, underscores, hyphens, or dots
  checkpoint        Path to the inference checkpoint file

options:
  -h, --help        show this help message and exit
  --files [FILES ...], -f [FILES ...]
                    Additional files to upload to the model directory on S3
  --overwrite, -o   Overwrite existing files
```

It extracts the metadata from the checkpoint and stores it in
`tests/integration/{model}/` together with an example config. Supporting
arrays are also extracted and uploaded to S3 under
`s3://ml-tests/test-data/samples/anemoi-integration-tests/inference/{model}/`.

I did not put this script in the CLI because this is a developer-only
script, and will create files to its relative location in the repo.

### RMI LAM
The model was trained on a O48 inner and O32 boundary region. Reduced
testing datasets were made available in the catalogue. Because the RMI
setup runs from a zarr dataset, a further-reduced 5 day version of the
test datasets was made locally to be used as input data. These were
zipped and provided to the above script as extra `--files` and uploaded
to S3.

The test configuration uses the facility of anemoi-datasets to download
zarr zips at runtime. Because the input datasets are tiny (~10MB), this
works well and is fast.

Note: when doing this, care needs to be taken that there is no
subdirectory in the zip: the root path of the zarr should be the root
path of the zip. This was done with: `(cd
aifs-ea-an-oper-0001-mars-o48-2020-6h.zarr && zip -r
../aifs-ea-an-oper-0001-mars-o48-2020-6h.zarr.zip .)`

Because the mock integration test model now passes prognostic inputs to
the output, the new `check_lam_with_xarray` check compares the values in
the output file to a reference dataset (the input). The test config is
self-explanatory: we pass the same dataset used for the inner region to
the check, which in turn compares the value of each prognostic variable.
We also check that the number of grid points matches the mask in the
checkpoint.

## What issue or task does this change relate to?
closes #272
#273

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)


<!-- readthedocs-preview anemoi-inference start -->
----
📚 Documentation preview 📚:
https://anemoi-inference--308.org.readthedocs.build/en/308/

<!-- readthedocs-preview anemoi-inference end -->