fix(parallel runner): Support base class with main_argument decorator (#333)
## Description
If the `main_argument` decorator is used on a class and subsequently
stored in the registry, the registered object becomes a function. In
most cases this is not a problem, except in the parallel runner which
needs the original class of the base runner in the registry, as
described in #309.

Refactor this so the wrapping is done with a class instead of a
function.

The decorator is not used on functions anywhere, and I don't think there
is a use case to do so. This removes support for the decorator to be
used on functions, and only support classes.

## What issue or task does this change relate to?
fixes #309 

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)