fix(metadata): Patching when receiving entry is not a dict (#338)
## Description
Fixing an edge case in the `external_graph` runner when used with a LAM
model. When using this runner with the `graph_dataset` option, it will
patch the `dataloader.dataset` entry:


https://github.com/ecmwf/anemoi-inference/blob/196aa55a57204c4bcb823e7b6546d79505f5eaee/src/anemoi/inference/runners/external_graph.py#L189-L199

But if the existing `dataloader.dataset` entry in the metadata happens
to be a list of datasets, the patching crashes. When patching an
existing entry, the patch function did not check if the receiving entry
is also a dictionary. In the case of a list on the receiving side, it
would try to index the list with the incoming dictionary keys. This PR
fixes that.

Also adds a test for this function.


***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)