feat(grib): Improve template manager (#383)
## Description
Improve the grib template manager in a backwards compatible way.

- [x] Improve documentation on how users can provide their own templates
- [x] For the `samples` provider, add the option for the sample index to
be specified directly in the inference config
- [x] For the `file` provider, add an option for the file to contain
many parameters, and do a lookup based on the param name.
- [x] In the case of grib input, the template manager automatically
injects templates found in the input grib at the highest priority.
Refactor this so that this becomes its own standalone template provider
called `input`, this gives more flexibility to enable/disable
combinations of template providers and set the priority..
- [x] `input` provider fallback mode: force an output variable to use
one of the input variables
- [x] Improve logging of the template manager. Right now it only logs in
the case of errors. Also log in the case of success so it's easier to
see which templates are being used from which provider.
- [x] tests for everything, including backwards compatibility

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)


<!-- readthedocs-preview anemoi-inference start -->
----
📚 Documentation preview 📚:
https://anemoi-inference--383.org.readthedocs.build/en/383/

<!-- readthedocs-preview anemoi-inference end -->