feat(grib): Control behaviour of negative output steps (#392)
## Description
Add a new option to the grib output:
```
        negative_step_mode : Literal["error", "write", "skip"], optional
            What to do when writing a variable that has a base time before the forecast base time.
            This can happen when the initial conditions contain an accumulated variable, or a variable period is longer than the model step time.
            In all cases a warning will be shown.
            - `error`: (default) raise an exception
            - `write`: write the variable as normal
            - `skip`: skip writing the variable
```

When the initial state contains accumulated or average variables, or if
variables have a period longer than the model timestep, the output time
step can become "negative": the base time of the variable is before the
base time of the forecast.

This flag allows the user to choose what to do when this happens. 


***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: b8raoult <53792887+b8raoult@users.noreply.github.com>