feat(interpolator): Default mars request from forecast (#397)
## Description
When running the interpolator from data in MARS (like we do internally
in prepml), the `anemoi-inference retrieve` command can be used to
generate the initial conditions retrieval request to get the input grib.
This command expects an inference config with a certain runner, and the
runner has a chance to patch the data request according to its needs.

For the interpolator, it makes sense that the default retrieval request
is a forecast and not analysis. This PR sets the default to retrieve
from ECMWF operational forecast, with the length of the forecast decided
by the `lead_time`.

@frazane @OpheliaMiralles This PR should not impact you or other
external users, as the `patch_data_request` function is not used
anywhere else other than the retrieve command. But could you still check
if this is okay for you?

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)